refactor: add osutil package
